requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pillow==10.0.0
sqlalchemy
psycopg2-binary
alembic
boto3
pydantic-settings
python-dotenv
